EDGORDERS-57 migrated vertx to the v4.3.1

+++ src/main/java/org/folio/edge/orders/utils/OrdersOkapiClientFactory.java
@@ -6,7 +6,7 @@ import io.vertx.core.Vertx;
 
 public class OrdersOkapiClientFactory extends OkapiClientFactory {
 
-  public OrdersOkapiClientFactory(Vertx vertx, String okapiURL, long reqTimeoutMs) {
+  public OrdersOkapiClientFactory(Vertx vertx, String okapiURL, int reqTimeoutMs) {
     super(vertx, okapiURL, reqTimeoutMs);
   }
